+++ __tests__/__integration__/imperative/__tests__/__integration__/cli/config/convert-profiles/__scripts__/delete_profiles_secured_and_base.sh
@@ -1,0 +1,10 @@
+#!/bin/bash
+
+imperative-test-cli profiles delete secured test
+if [ $? -gt 0 ]
+then
+    exit $?
+fi
+
+imperative-test-cli profiles delete base test
+exit $?