+++ packages/config/src/doc/IConfigConvertResult.ts
@@ -9,5 +9,14 @@
 *
 */
 
-const open = jest.genMockFromModule('open');
-module.exports = open;
+import { IConfig } from "./IConfig";
+
+export interface IConfigConvertResult {
+    config: IConfig;
+    profilesConverted: { [key: string]: string[] };
+    profilesFailed: {
+        name?: string;
+        type: string;
+        error: Error;
+    }[];
+}