+++ packages/imperative/src/config/cmd/auto-init/AutoInitConstants.ts
@@ -46,7 +46,7 @@ export class AutoInitConstants {
         aliases: ["dr", "dry"],
         description: "Display the outcome of the initialization without saving",
         type: "boolean",
-        defaultValue: false,
+        conflictsWith: ["edit"],
         group: AutoInitConstants.AUTO_INIT_OPTION_GROUP
     };
 
@@ -56,9 +56,9 @@ export class AutoInitConstants {
     public static AUTO_INIT_OPTION_EDIT: ICommandOptionDefinition = {
         name: "edit",
         aliases: ["e"],
-        description: "Open an editor for the configuration",
+        description: "Open in editor after initializing the configuration",
         type: "boolean",
-        defaultValue: false,
+        conflictsWith: ["dry-run"],
         group: AutoInitConstants.AUTO_INIT_OPTION_GROUP
     };